# docs/swagger.yaml
basePath: /
definitions:
  model.AddPolyclinicRequest:
    description: Hastaneye poliklinik ekleme verisi
    properties:
      floor:
        description: Kat numarası
        example: 2
        type: integer
      polyclinic_type_id:
        description: Master data'dan seçilen tür
        example: 1
        type: integer
      room_number:
        description: Oda numarası
        example: 205
        type: integer
    required:
    - floor
    - polyclinic_type_id
    - room_number
    type: object
...
  model.District:
    description: İlçe bilgileri
    properties:
      name:
        description: İlçe adı
        example: Beşiktaş
        type: string
      province:
        allOf:
        - $ref: '#/definitions/model.Province'
        description: İl bilgisi
      province_id:
        description: Bağlı olduğu il
        example: 1
        type: integer
    required:
    - name
    - province_id
    type: object
  model.Hospital:
    description: Hastane bilgileri
    properties:
      address_detail:
        description: Açık adres
        example: Beşiktaş Caddesi No:123
        type: string
      district:
        allOf:
        - $ref: '#/definitions/model.District'
        description: İlçe bilgisi
      district_id:
        description: İlçe ID
        example: 1
        type: integer
      email:
        description: E-posta adresi
        example: info@acibadem.com
        type: string
      name:
        description: Hastane adı
        example: Acıbadem Hastanesi
        type: string
      phone:
        description: Telefon numarası
        example: "02121234567"
        type: string
      province:
        allOf:
        - $ref: '#/definitions/model.Province'
        description: İlişkiler
      province_id:
        description: İl ID
        example: 1
        type: integer
      tax_id:
        description: Vergi kimlik numarası
        example: "1234567890"
        type: string
      users:
        description: Hastane kullanıcıları
        items:
          $ref: '#/definitions/model.User'
        type: array
    required:
    - address_detail
    - district_id
    - email
    - name
    - phone
    - province_id
    - tax_id
    type: object
  model.HospitalPolyclinic:
    description: Hastane poliklinik bilgileri
    properties:
      floor:
        description: Kat numarası
        example: 2
        type: integer
      hospital:
        allOf:
        - $ref: '#/definitions/model.Hospital'
        description: İlişkiler
      hospital_id:
        description: Hangi hastane
        example: 1
        type: integer
      is_active:
        description: Aktif mi?
        example: true
        type: boolean
      polyclinic_type:
        $ref: '#/definitions/model.PolyclinicType'
      polyclinic_type_id:
        description: Poliklinik türü
        example: 1
        type: integer
      room_number:
        description: Oda numarası
        example: 205
        type: integer
      staff:
        description: Bu poliklinikte çalışan personel
        items:
          $ref: '#/definitions/model.Staff'
        type: array
    required:
    - floor
    - hospital_id
    - polyclinic_type_id
    - room_number
    type: object
  model.HospitalPolyclinicSummary:
    description: Hastane poliklinik özet bilgileri
    properties:
      floor:
        example: 2
        type: integer
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      polyclinic_type_name:
        example: Kardiyoloji
        type: string
      room_number:
        example: 205
        type: integer
      staff_by_job_group:
        description: Meslek grubuna göre personel sayıları
        items:
          $ref: '#/definitions/model.StaffCountByGroup'
        type: array
      total_staff_count:
        description: Toplam personel sayısı
        example: 10
        type: integer
    type: object
...
  model.JobGroup:
    description: Meslek grubu bilgileri
    properties:
      job_titles:
        description: Bu gruba ait unvanlar
        items:
          $ref: '#/definitions/model.JobTitle'
        type: array
      name:
        description: Meslek grubu adı
        example: Doktor
        type: string
    required:
    - name
    type: object
  model.JobTitle:
    description: Unvan bilgileri
    properties:
      is_unique:
        description: 'Bu unvandan hastanede sadece 1 tane olabilir mi? (örn: Başhekim)'
        example: false
        type: boolean
      job_group:
        allOf:
        - $ref: '#/definitions/model.JobGroup'
        description: Meslek grubu bilgisi
      job_group_id:
        description: Bağlı olduğu meslek grubu
        example: 1
        type: integer
      name:
        description: Unvan adı
        example: Uzman Doktor
        type: string
    required:
    - job_group_id
    - name
    type: object
...
  model.PolyclinicType:
    description: Poliklinik türü bilgileri (master data)
    properties:
      description:
        description: Açıklama
        example: Kalp ve damar hastalıkları
        type: string
      name:
        description: Poliklinik türü adı
        example: Kardiyoloji
        type: string
    required:
    - name
    type: object
  model.Province:
    description: İl bilgileri
    properties:
      districts:
        description: İlçeler
        items:
          $ref: '#/definitions/model.District'
        type: array
      name:
        description: İl adı
        example: İstanbul
        type: string
    required:
    - name
    type: object
  model.ResetPasswordConfirm:
    description: Şifre sıfırlama onayı
    properties:
      code:
        description: Doğrulama kodu
        example: "123456"
        type: string
      confirm_password:
        description: Şifre tekrarı
        example: yeni123456
        type: string
      new_password:
        description: Yeni şifre
        example: yeni123456
        minLength: 6
        type: string
      phone:
        description: Telefon numarası
        example: "05551234567"
        type: string
    required:
    - code
    - confirm_password
    - new_password
    - phone
    type: object
  model.ResetPasswordRequest:
    description: Şifre sıfırlama talebi
    properties:
      phone:
        description: Telefon numarası
        example: "05551234567"
        type: string
    required:
    - phone
    type: object
  model.Staff:
    description: Hastane personel bilgileri
    properties:
      first_name:
        description: Ad
        example: Dr. Mehmet
        type: string
      hospital:
        allOf:
        - $ref: '#/definitions/model.Hospital'
        description: İlişkiler
      hospital_id:
        description: Hangi hastane
        example: 1
        type: integer
      is_active:
        description: Aktif mi?
        example: true
        type: boolean
      job_group:
        $ref: '#/definitions/model.JobGroup'
      job_group_id:
        description: Meslek grubu
        example: 1
        type: integer
      job_title:
        $ref: '#/definitions/model.JobTitle'
      job_title_id:
        description: Unvan
        example: 1
        type: integer
      last_name:
        description: Soyad
        example: Özkan
        type: string
      phone:
        description: Telefon
        example: "05559876543"
        type: string
      polyclinic:
        $ref: '#/definitions/model.HospitalPolyclinic'
      polyclinic_id:
        description: Bağlı olduğu poliklinik (nullable - güvenlik gibi genel personel
          için)
        example: 1
        type: integer
      tc:
        description: TC Kimlik No
        example: "98765432101"
        type: string
      work_days:
        description: 'Çalışma günleri (JSON array: 1=Pazartesi, 7=Pazar)'
        example: '[1,2,3,4,5]'
        type: string
    required:
    - first_name
    - hospital_id
    - job_group_id
    - job_title_id
    - last_name
    - phone
    - tc
    - work_days
    type: object
  model.StaffCountByGroup:
    description: Meslek grubuna göre personel sayısı
    properties:
      count:
        example: 7
        type: integer
      job_group_name:
        example: Doktor
        type: string
    type: object
...
  model.UpdatePolyclinicRequest:
    description: Hastane poliklinik güncelleme verisi
    properties:
      floor:
        description: Kat numarası
        example: 3
        type: integer
      is_active:
        description: Aktif mi?
        example: true
        type: boolean
      room_number:
        description: Oda numarası
        example: 301
        type: integer
    required:
    - floor
    - room_number
    type: object
...
  model.User:
    description: Hastane kullanıcı bilgileri
    properties:
      created_by:
        description: Kim tarafından eklendi (nullable - ilk user için)
        example: 1
        type: integer
      creator:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Oluşturan kullanıcı
      email:
        description: E-posta adresi
        example: ahmet.yilmaz@example.com
        type: string
      first_name:
        description: Ad
        example: Ahmet
        type: string
      hospital:
        allOf:
        - $ref: '#/definitions/model.Hospital'
        description: İlişkiler
      hospital_id:
        description: Hangi hastaneye ait
        example: 1
        type: integer
      is_active:
        description: Aktif mi?
        example: true
        type: boolean
      last_name:
        description: Soyad
        example: Yılmaz
        type: string
      password:
        description: Şifre (JSON'dan okuyabilir ama response'da göstermez)
        example: "123456"
        minLength: 6
        type: string
      phone:
        description: Telefon numarası
        example: "05551234567"
        type: string
      role:
        description: 'Rol: "yetkili" veya "çalışan"'
        enum:
        - yetkili
        - çalışan
        example: çalışan
        type: string
      tc:
        description: Türkiye Cumhuriyeti Kimlik Numarası
        example: "12345678901"
        type: string
    required:
    - email
    - first_name
    - hospital_id
    - last_name
    - password
    - phone
    - role
    - tc
    type: object
host: localhost:8080
info:
  contact: {}
  description: Bu API VatanSoft staj giriş projesi için yazılmıştır.
  title: Hastane Takip API
  version: "1.0"
paths:
...
  /hospital/polyclinics:
    get:
      description: Hastaneye ait poliklinikleri personel sayılarıyla listeler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.HospitalPolyclinicSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hastane polikliniklerini getir
      tags:
      - Polyclinic
    post:
      consumes:
      - application/json
      description: Seçilen poliklinik türünü hastaneye ekler
      parameters:
      - description: Poliklinik ekleme verisi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddPolyclinicRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HospitalPolyclinic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hastaneye poliklinik ekle
      tags:
      - Polyclinic
  /hospital/polyclinics/{id}:
    delete:
      description: Hastane poliklinik siler
      parameters:
      - description: Poliklinik ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hastane poliklinik sil
      tags:
      - Polyclinic
    put:
      consumes:
      - application/json
      description: Hastane poliklinik bilgilerini günceller
      parameters:
      - description: Poliklinik ID
        in: path
        name: id
        required: true
        type: integer
      - description: Güncelleme verisi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePolyclinicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HospitalPolyclinic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hastane poliklinik güncelle
      tags:
      - Polyclinic
...
  /polyclinic-types:
    get:
      description: Master data'dan poliklinik türlerini getirir (dropdown için)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PolyclinicType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Poliklinik türleri listesi
      tags:
      - Polyclinic
...
  /reset-password/confirm:
    post:
      consumes:
      - application/json
      description: Kod doğruysa yeni şifre veritabanına yazılır
      parameters:
      - description: Yeni şifre bilgileri
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Kodla şifreyi sıfırla
      tags:
      - Auth
  /reset-password/request:
    post:
      consumes:
      - application/json
      description: Telefon numarasına geçici şifre sıfırlama kodu üretir
      parameters:
      - description: Telefon numarası
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Şifre sıfırlama kodu gönder
      tags:
      - Auth
swagger: "2.0"